refactor: separate pagination controls to its own component
