+++ web/components/templates/hql/QueryResult.tsx
@@ -26,9 +26,10 @@ import { $JAWN_API } from "@/lib/clients/jawn";
 import { CellContext } from "@tanstack/react-table";
 import { components } from "@/lib/clients/jawnTypes/public";
 import useNotification from "@/components/shared/notification/useNotification";
-import { CircleCheckBig, CircleDashed, ChevronLeft, ChevronRight, ChevronsLeft, ChevronsRight } from "lucide-react";
+import { CircleCheckBig, CircleDashed } from "lucide-react";
 import { Progress } from "@/components/ui/progress";
 import { HqlErrorDisplay } from "./HqlErrorDisplay";
+import { PaginationControls } from "./PaginationControls";
 
 interface QueryResultProps {
   sql: string;
@@ -336,104 +337,5 @@ function ExportButton({ sql }: ExportButtonProps) {
     </TooltipProvider>
   );
 }
-
-interface PaginationControlsProps {
-  currentPage: number;
-  totalPages: number;
-  rowsPerPage: number;
-  totalRows: number;
-  startRow: number;
-  endRow: number;
-  onPageChange?: (page: number) => void;
-  onRowsPerPageChange?: (rows: number) => void;
-  isBottom?: boolean;
-}
-
-const PaginationControls = ({
-  currentPage,
-  totalPages,
-  rowsPerPage,
-  totalRows,
-  startRow,
-  endRow,
-  onPageChange,
-  onRowsPerPageChange,
-  isBottom = false,
-}: PaginationControlsProps) => {
-  const canGoPrevious = currentPage > 1;
-  const canGoNext = currentPage < totalPages;
-
-  return (
-    <div className={clsx(
-      "flex items-center justify-between border-tremor-brand-subtle bg-background px-4 py-2",
-      isBottom ? "border-t" : "border-b"
-    )}>
-      <div className="flex items-center gap-4">
-        <span className="text-xs text-muted-foreground">
-          Showing {startRow} to {endRow} of {totalRows} rows
-        </span>
-        
-        <Select
-          value={rowsPerPage.toString()}
-          onValueChange={(value) => onRowsPerPageChange?.(parseInt(value))}
-        >
-          <SelectTrigger className="h-8 w-[100px]">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="25">25 rows</SelectItem>
-            <SelectItem value="50">50 rows</SelectItem>
-            <SelectItem value="100">100 rows</SelectItem>
-            <SelectItem value="200">200 rows</SelectItem>
-          </SelectContent>
-        </Select>
-      </div>
-
-      <div className="flex items-center gap-2">
-        <Button
-          variant="ghost"
-          size="sm"
-          onClick={() => onPageChange?.(1)}
-          disabled={!canGoPrevious}
-          className="h-8 w-8 p-0"
-        >
-          <ChevronsLeft className="h-4 w-4" />
-        </Button>
-        <Button
-          variant="ghost"
-          size="sm"
-          onClick={() => onPageChange?.(currentPage - 1)}
-          disabled={!canGoPrevious}
-          className="h-8 w-8 p-0"
-        >
-          <ChevronLeft className="h-4 w-4" />
-        </Button>
-        
-        <span className="text-xs text-muted-foreground">
-          Page {currentPage} of {totalPages}
-        </span>
-        
-        <Button
-          variant="ghost"
-          size="sm"
-          onClick={() => onPageChange?.(currentPage + 1)}
-          disabled={!canGoNext}
-          className="h-8 w-8 p-0"
-        >
-          <ChevronRight className="h-4 w-4" />
-        </Button>
-        <Button
-          variant="ghost"
-          size="sm"
-          onClick={() => onPageChange?.(totalPages)}
-          disabled={!canGoNext}
-          className="h-8 w-8 p-0"
-        >
-          <ChevronsRight className="h-4 w-4" />
-        </Button>
-      </div>
-    </div>
-  );
-};
 
 export default QueryResult;
